requirements.txt
open-clip-torch==2.20.0
timm==0.9.16
# torch==2.1.2
# torchvision==0.16.2
numpy==1.26.4
Pillow==10.3.0
tqdm==4.66.5
ftfy==6.1.1
regex==2024.9.11
thop==0.1.1.post2209072238
einops==0.8.1
huggingface_hub==0.23.5
transformers==4.40.2
